data: fix #1318 type mismatch, enforce collab-silver contract, add dbt-coverage gate + CI
Fixes the NO_COMMON_TYPE (Float64 vs Int64) build failure in
class_collab_document_activity: visited_page_count is pinned to Nullable(Float64)
on both M365 halves (OneDrive cast NULL; SharePoint cast through).

Enables an ENFORCED dbt contract on class_collab_document_activity (every column
declared with its data_type from the live ClickHouse catalog; on_schema_change:
fail) so column/type drift fails the build, not production. First model on the
path to full contract coverage.

scripts/ci/dbt_coverage.py: walks the compiled manifest and flags metric-path
models (bronze-promoted -> silver) lacking an enforced contract or
not_null+unique on their key column, plus bronze source-freshness coverage.
Validated locally with dbt-core 1.10 + dbt-clickhouse 1.9 (dbt parse).

.github/workflows/data-contracts.yml: runs the gate in PR CI via dbt parse only
(no warehouse/credentials), report-only, uploading the per-model gap list.

Wording is neutral throughout (user-facing terms only).

Closes #1318. Related: #1319, #1321, #1326.

+++ src/ingestion/connectors/collaboration/m365/dbt/m365__collab_document_activity_onedrive.sql
@@ -28,7 +28,10 @@ SELECT
     syncedFileCount AS synced_count,
     sharedInternallyFileCount AS shared_internally_count,
     sharedExternallyFileCount AS shared_externally_count,
-    CAST(NULL AS Nullable(Int64)) AS visited_page_count,
+    -- OneDrive has no page-visit metric. Pinned to Nullable(Float64) to match
+    -- the SharePoint half: the silver tag-union (class_collab_document_activity)
+    -- fails with NO_COMMON_TYPE if the two halves disagree (Int64 vs Float64).
+    CAST(NULL AS Nullable(Float64)) AS visited_page_count,
     reportPeriod AS report_period,
     now() AS collected_at,
     'insight_m365' AS data_source,
